backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
groq==0.13.0
fastembed==0.5.2
qdrant-client==1.12.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sentence-transformers==3.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
aiofiles==24.1.0
cryptography==43.0.3
weasyprint==62.0
markdown2==2.4.12
EbookLib==0.18
pytest==8.3.2
pytest-mock==3.14.0
pytest-asyncio==0.23.8
PyYAML==6.0.1
slowapi==0.1.9
itsdangerous==2.2.0